third_party/blink/web_tests/external/wpt/css/css-values/calc-sibling-function-in-shadow-dom-expected.txt

This is a testharness.js-based test.
[FAIL] originating element in Shadow DOM
  assert_equals: expected "2" but got "4"
Harness: the test ran to completion.
